oleacc: pass structs on stack for 32-bit calling convention
The original author of this code never tested it.

Fixes #95
