+++ oleacc.go
@@ -218,24 +218,6 @@ func (obj *IAccPropServices) Release() uint32 {
 	return uint32(ret)
 }
 
-// SetPropValue identifies the accessible element to be annotated, specify the property to be annotated, and provide a new value for that property.
-// If server developers know the HWND of the accessible element they want to annotate, they can use one of the following methods: SetHwndPropStr, SetHwndProp, or SetHwndPropServer
-func (obj *IAccPropServices) SetPropValue(idString []byte, idProp *MSAAPROPID, v *VARIANT) HRESULT {
-	var idStringPtr unsafe.Pointer
-	idStringLen := len(idString)
-	if idStringLen != 0 {
-		idStringPtr = unsafe.Pointer(&idString[0])
-	}
-	ret, _, _ := syscall.Syscall6(obj.LpVtbl.SetPropValue, 5,
-		uintptr(unsafe.Pointer(obj)),
-		uintptr(idStringPtr),
-		uintptr(idStringLen),
-		uintptr(unsafe.Pointer(idProp)),
-		uintptr(unsafe.Pointer(v)),
-		0)
-	return HRESULT(ret)
-}
-
 // SetPropServer specifies a callback object to be used to annotate an array of properties for the accessible element. You can also specify whether the annotation is to be applied to this accessible element or to the element and its children. This method is used for server annotation.
 // If server developers know the HWND of the accessible element they want to annotate, they can use SetHwndPropServer.
 func (obj *IAccPropServices) SetPropServer(idString []byte, idProps []MSAAPROPID, server *IAccPropServer, annoScope AnnoScope) HRESULT {
@@ -282,34 +264,6 @@ func (obj *IAccPropServices) ClearProps(idString []byte, idProps []MSAAPROPID) H
 		uintptr(idPropsPtr),
 		uintptr(idPropsLen),
 		0)
-	return HRESULT(ret)
-}
-
-// SetHwndProp wraps SetPropValue, providing a convenient entry point for callers who are annotating HWND-based accessible elements. If the new value is a string, you can use SetHwndPropStr instead.
-func (obj *IAccPropServices) SetHwndProp(hwnd HWND, idObject int32, idChild uint32, idProp *MSAAPROPID, v *VARIANT) HRESULT {
-	ret, _, _ := syscall.Syscall6(obj.LpVtbl.SetHwndProp, 6,
-		uintptr(unsafe.Pointer(obj)),
-		uintptr(hwnd),
-		uintptr(idObject),
-		uintptr(idChild),
-		uintptr(unsafe.Pointer(idProp)),
-		uintptr(unsafe.Pointer(v)))
-	return HRESULT(ret)
-}
-
-// SetHwndPropStr wraps SetPropValue, providing a more convenient entry point for callers who are annotating HWND-based accessible elements.
-func (obj *IAccPropServices) SetHwndPropStr(hwnd HWND, idObject int32, idChild uint32, idProp *MSAAPROPID, str string) HRESULT {
-	str16, err := syscall.UTF16PtrFromString(str)
-	if err != nil {
-		return -((E_INVALIDARG ^ 0xFFFFFFFF) + 1)
-	}
-	ret, _, _ := syscall.Syscall6(obj.LpVtbl.SetHwndPropStr, 6,
-		uintptr(unsafe.Pointer(obj)),
-		uintptr(hwnd),
-		uintptr(idObject),
-		uintptr(idChild),
-		uintptr(unsafe.Pointer(idProp)),
-		uintptr(unsafe.Pointer(str16)))
 	return HRESULT(ret)
 }
 
@@ -387,34 +341,6 @@ func (obj *IAccPropServices) DecomposeHwndIdentityString(idString []byte) (hr HR
 		uintptr(unsafe.Pointer(&idChild)))
 	hr = HRESULT(ret)
 	return
-}
-
-// SetHmenuProp wraps SetPropValue, providing a convenient entry point for callers who are annotating HMENU-based accessible elements. If the new value is a string, you can use IAccPropServices::SetHmenuPropStr instead.
-func (obj *IAccPropServices) SetHmenuProp(hmenu HMENU, idChild uint32, idProp *MSAAPROPID, v *VARIANT) HRESULT {
-	ret, _, _ := syscall.Syscall6(obj.LpVtbl.SetHmenuProp, 5,
-		uintptr(unsafe.Pointer(obj)),
-		uintptr(hmenu),
-		uintptr(idChild),
-		uintptr(unsafe.Pointer(idProp)),
-		uintptr(unsafe.Pointer(v)),
-		0)
-	return HRESULT(ret)
-}
-
-// SetHmenuPropStr wraps SetPropValue, providing a more convenient entry point for callers who are annotating HMENU-based accessible elements.
-func (obj *IAccPropServices) SetHmenuPropStr(hmenu HMENU, idChild uint32, idProp *MSAAPROPID, str string) HRESULT {
-	str16, err := syscall.UTF16PtrFromString(str)
-	if err != nil {
-		return -((E_INVALIDARG ^ 0xFFFFFFFF) + 1)
-	}
-	ret, _, _ := syscall.Syscall6(obj.LpVtbl.SetHmenuPropStr, 5,
-		uintptr(unsafe.Pointer(obj)),
-		uintptr(hmenu),
-		uintptr(idChild),
-		uintptr(unsafe.Pointer(idProp)),
-		uintptr(unsafe.Pointer(str16)),
-		0)
-	return HRESULT(ret)
 }
 
 // SetHmenuPropServer wraps SetPropServer, providing a convenient entry point for callers who are annotating HMENU-based accessible elements.
